test: Make Hugr valid in replace_types::op_to_call (#2732)
The pre-ReplaceTypes hugr contained a FuncDefn nested within a DFG,
built by manually inserting into a built Hugr and not revalidating
(which would have complained about the non-top-level FuncDefn)...and we
weren't validating *after* ReplaceTypes either.

So build the thing properly using `module_root_builder` and validate,
which strengthens the test a bit.