refactor(linking.rs)!: (tiny) avoid type_complexity (#2721)
This is breaking in that if you called
`insert_link_view_by_node::<SomeHugrViewType>` you would need to change
that to instead specify the Node type of the view. That is a small
enough breakage that semver-checks doesn't spot it....but I think we can
roll into the 0.25.0 breaking release.

BREAKING CHANGE: `insert_link_view_by_node` parametric over type of Node
not HugrView