chore: Use working syntax for nextest install check in justfile setup (#2743)
The nextest install check during `just setup` was broken since it used a
just recipe invocation syntax that is not valid. The `setup` recipe has
been updated to use a "post" command, which is equivalent to running the
named recipe as a last command in the parent recipe. See
https://github.com/casey/just/discussions/818#discussioncomment-11854830
for more information.

Discovered during Hugr setup.

Co-authored-by: Douglas Wilson <douglas.wilson@quantinuum.com>