refactor!: Delete ValueArray (#2760)
Joint PR with @tatiana-s
closes #2398

Removes ValueArray and conversion to Array.

The handlers for linearizing (non-linear code using) Array remain, as
these do not actually depend on ValueArray, and we still need the
`num_outports==0` case for Guppy's `drop` op, so I'm retaining the code
for arbitrary #outports even though this is not used "in production".

BREAKING CHANGE: ValueArray is gone, as is LinearizeArrayPass

---------

Co-authored-by: Tatiana S <tatiana.sedelnikov@quantinuum.com>