fix: two problems in array linearization from #2749 (#2779)
#2749 suffered from poor test coverage, and although I thought I had run
the guppy test suite using a modified `selene-hugr-qis-compiler` I
failed to reveal the first bug here (and the rest were not covered by
the guppy test suite of the time).

* Avoid generating polymorphic `__unwrap` and `__mk_none` helper
functions; ReplaceTypes generally runs after monomorphization so this
avoids needing to rerun the latter.
* `copy_scan_func_name` did not depend on the size of the array, yet the
generated helper function took and returned an array (of options, plus
an index) - this means linking raises a conflicting-signature error if
you are linearizing copies of two different-size arrays of the same
element type
* add test of the latter